linting

+++ packages/server/src/routes/transactions.ts
@@ -1,10 +1,10 @@
-import * as Express from "express";
-import { PostgresDB } from "../types";
-import { getTransactions } from "../queries/getTransactions";
-import { createTransaction } from "../queries/createTransaction";
-import { getBudgetById } from "../queries/getBudgetById";
-import { updateTransaction } from "../queries/updateTransaction";
-import { deleteTransaction } from "../queries/deleteTransaction";
+import * as Express from 'express';
+import { PostgresDB } from '../types';
+import { getTransactions } from '../queries/getTransactions';
+import { createTransaction } from '../queries/createTransaction';
+import { getBudgetById } from '../queries/getBudgetById';
+import { updateTransaction } from '../queries/updateTransaction';
+import { deleteTransaction } from '../queries/deleteTransaction';
 
 interface Modified {
   isNew?: boolean;
@@ -13,14 +13,11 @@ interface Modified {
 
 interface SaveTransactionsBody {
   id: number;
-  transactions: (
-    | (Budget.Transaction & Modified)
-    | { id: number; isDeleted: true }
-  )[];
+  transactions: ((Budget.Transaction & Modified) | { id: number; isDeleted: true })[];
 }
 
 export const registerRoutes = (app: Express.Application, db: PostgresDB) => {
-  app.get("/transactions", async (req, res) => {
+  app.get('/transactions', async (req, res) => {
     try {
       const transactions = await getTransactions(db);
       return res.json(transactions);
@@ -33,7 +30,7 @@ export const registerRoutes = (app: Express.Application, db: PostgresDB) => {
     }
   });
 
-  app.post("/transaction", async (req, res) => {
+  app.post('/transaction', async (req, res) => {
     try {
       const id = await createTransaction(db, req.body);
       return res.json(id);
@@ -46,21 +43,19 @@ export const registerRoutes = (app: Express.Application, db: PostgresDB) => {
     }
   });
 
-  app.post("/transactions", async (req, res) => {
+  app.post('/transactions', async (req, res) => {
     try {
       const { id, transactions } = req.body as SaveTransactionsBody;
 
       await Promise.all(
         transactions.map(transaction => {
-          if ("isDeleted" in transaction) {
-            if (transaction.isDeleted)
-              return deleteTransaction(db, transaction.id);
+          if ('isDeleted' in transaction) {
+            if (transaction.isDeleted) return deleteTransaction(db, transaction.id);
           } else {
             if (transaction.isNew) return createTransaction(db, transaction);
-            if (transaction.isUpdated)
-              return updateTransaction(db, transaction);
+            if (transaction.isUpdated) return updateTransaction(db, transaction);
           }
-        })
+        }),
       );
 
       const budget = await getBudgetById(db, String(id));
